AI-SocialMedia-System-main/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0

# Google APIs
googlemaps==4.10.0
google-api-python-client==2.108.0

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0

# Web scraping and APIs
selenium==4.15.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0


